Merge V3 model updates

+++ R/download_models.R
@@ -16,7 +16,7 @@
 #' 
 #' @export
 #' 
-download_models <- function(models = "species_v2")
+download_models <- function(models)
   {
 
   # set cache destination
@@ -31,20 +31,24 @@ download_models <- function(models = "species_v2")
     model_i <- models[i]
     
     # define latest model generation - manually update with each deployment
-    latest <- "v2"
+    if(model_i == "species") {
+      latest <- "v3"
+    } else{
+      latest <- "v2"
+    }
     if(model_i %in% c('general', 'family', 'species', 'pig_only')){
       model_i <- paste(model_i, latest, sep="_")}
     
     # define file path depending on model version
     if(grepl("general", model_i, ignore.case = TRUE)) {
       if(grepl("general_v1", model_i, ignore.case = TRUE)){
-        path <- file.path(cache_path, fs::path_file("general_v1_1.zip"))
-        folder <- file.path(cache_path, fs::path_file("general_v1"))
+        path <- fs::path(cache_path, "general_v1_1", ext="zip")
+        folder <- fs::path(cache_path, "general_v1")
         #agdata <- "https://agdatacommons.nal.usda.gov/ndownloader/files/44576746"
         agdata <- "https://bit.ly/4c2Ykz2"
       } else {
-        path <- file.path(cache_path, fs::path_file("general_v2.zip"))
-        folder <- file.path(cache_path, fs::path_file("general_v2"))
+        path <- fs::path(cache_path, "general_v2", ext="zip")
+        folder <- fs::path(cache_path, "general_v2")
         #agdata <- "https://agdatacommons.nal.usda.gov/ndownloader/files/44576752"
         agdata <- "https://bit.ly/4c42iHt"
       }
@@ -52,47 +56,51 @@ download_models <- function(models = "species_v2")
     }
     if(grepl("family", model_i, ignore.case = TRUE)) {
       if(grepl("family_v1", model_i, ignore.case = TRUE)){
-        path <- file.path(cache_path, fs::path_file("family_v1_1.zip"))
-        folder <- file.path(cache_path, fs::path_file("family_v1"))
+        path <- fs::path(cache_path, "family_v1_1", ext="zip")
+        folder <- fs::path(cache_path, "family_v1")
         #agdata <- "https://agdatacommons.nal.usda.gov/ndownloader/files/44576266"
         agdata <- "https://bit.ly/438ZCUU"
       } else {
-        path <- file.path(cache_path, fs::path_file("family_v2.zip"))
-        folder <- file.path(cache_path, fs::path_file("family_v2"))
+        path <- fs::path(cache_path, "family_v2", ext="zip")
+        folder <- fs::path(cache_path, "family_v2")
         #agdata <- "https://agdatacommons.nal.usda.gov/ndownloader/files/44576296"
         agdata <- "https://bit.ly/3T3Yhu1"
       }
       
     }
     if(grepl("pig", model_i, ignore.case = TRUE)) {
-      path <- file.path(cache_path, fs::path_file("pig_v1_0.zip"))
-      folder <- file.path(cache_path, fs::path_file("pig_v1"))
+      path <- fs::path(cache_path, "pig_v1_0", ext="zip")
+      folder <- fs::path(cache_path, "pig_v1")
       #agdata <- "https://agdatacommons.nal.usda.gov/ndownloader/files/44576767"
       agdata <- "https://bit.ly/438zJ7J"
     }
     if(grepl("species", model_i, ignore.case = TRUE)) {
       if(grepl("species_v1", model_i, ignore.case = TRUE)) {
-        path <- file.path(cache_path, fs::path_file("species_v1_1.zip"))
-        folder <- file.path(cache_path, fs::path_file("species_v1"))
+        path <- fs::path(cache_path, "species_v1_1", ext="zip")
+        folder <- fs::path(cache_path, "species_v1")
         #agdata <- "https://agdatacommons.nal.usda.gov/ndownloader/files/44576215"
         agdata <- "https://bit.ly/3T4srgM"
       }
-      else {
-        path <- file.path(cache_path, fs::path_file("species_v2_3.zip"))
-        folder <- file.path(cache_path, fs::path_file("species_v2"))
+      if(grepl("species_v2", model_i, ignore.case = TRUE)) {
+        path <- fs::path(cache_path, "species_v2_3", ext="zip")
+        folder <- fs::path(cache_path, "species_v2")
         #agdata <- "https://agdatacommons.nal.usda.gov/ndownloader/files/44576230"
         agdata <- "https://bit.ly/ctdv2sp"
+      } else {
+        path <- fs::path(cache_path, "species_v3", ext="zip")
+        folder <- fs::path(cache_path, "species_v3")
+        agdata <- 'https://bit.ly/ctdv3sp'
       }
       
     }
     
     # unzip the folder if it's in the cache
-    if(file.exists(path)){
+    if(fs::file_exists(path)){
       # unzip the folder
       utils::unzip(zipfile=path, exdir=cache_path)
     }
     
-    if (!dir.exists(folder)){
+    if (!fs::dir_exists(folder)){
       
       #print timeout message only once
       if(i == 1) {
@@ -103,12 +111,14 @@ download_models <- function(models = "species_v2")
       }
       
       # download and unzip model files
-      #utils::download.file(url=agdata, destfile=path, quiet=TRUE)
-      tryCatch(usethis::use_zip(agdata, destdir = cache_path), 
+      # utils::download.file(url=agdata, destfile=path, quiet=TRUE)
+      tryCatch(usethis::use_zip(agdata, destdir = cache_path, cleanup=T), 
                error = function(e) {utils::unzip(path, exdir = cache_path)})
       
-      # once we confirm download, remove zip file
-        file.remove(path)
+      # once we confirm download, remove zip file if it is still there
+        if(fs::file_exists(path)) {
+          file.remove(path)
+        }
     }
     
     cat(paste0("Model files for model version ", model_i, " are downloaded and stored.\n\n"))
